+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+- Update extracting trace from headers to work with new version of Lmc.Trace
 
 ## 16.0.0 - 2021-07-27
 - Add `Producer` functions